Return instructor_id as a number, not a string
instructor_id is BIGINT, and node-postgres returns BIGINT as a string, so the
session token and /auth/me were handing the client "58" rather than 58. Nothing
breaks today - the scoped query coerces it in SQL, and no comparison relies on
it - but any === against a numeric instructor id would silently be false, which
is a bad trap to leave in an authorization-adjacent field.

Coerced in signToken/publicUser, and cast with ::int in the /auth/me and
/settings/users selects.

Found while verifying the first real instructor account end to end.

+++ server/routes/settings.js
@@ -78,7 +78,7 @@ router.delete('/delegates/:id', async (req, res) => {
 router.get('/users', async (req, res) => {
   const { rows } = await pool.query(
     `SELECT u.id, u.name, u.initials, u.email, u.role, u.active, u.created_at,
-            u.instructor_id, i.name AS instructor_name
+            u.instructor_id::int, i.name AS instructor_name
        FROM users u
        LEFT JOIN instructors i ON i.id = u.instructor_id
       ORDER BY u.name`
